feat:#04 placing a destructuring assignment (also known as split assignment) in the constructor is not a bad approach. In fact, it can make your code cleaner and easier to read by reducing the amount of code you need to write.
